chore: update errorManager
